move out opencode logic for consistency
